third_party/blink/web_tests/fast/xsl/xslt-text-expected.txt

CONSOLE WARNING: crbug.com/435623334: This page uses XSLT, which being considered for removal from the web. If that happens, it is possible that this page will need to be updated to maintain functionality.

<head xmlns="http://www.w3.org/1999/xhtml"><title></title></head>
<body xmlns="http://www.w3.org/1999/xhtml">
<pre>CHARACTERS IN XSLT: &lt;&lt;&lt;&amp;тест&amp;&gt;&gt;&gt;
SOURCE XML: &lt;&lt;&lt;&amp;тест&amp;&gt;&gt;&gt;</pre>
</body>



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
CHARACTERS IN XSLT: <<<&тест&>>>
SOURCE XML: <<<&тест&>>>
